requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
openai==1.65.1
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.5
regex==2024.11.6
requests==2.32.3
rich==13.9.4
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
daemonize==2.5.0
PyYAML==6.0.2
ordered-set==4.1.0
